Move payload related code to separate directory
