+++ rowhammer_tester/payload/ddr4lib.py
@@ -1,11 +1,12 @@
 import math
 import collections
-import payload_ddr3_pb2
 
-Timing = payload_ddr3_pb2.Timing
-Opcode = payload_ddr3_pb2.Opcode
-Instr = payload_ddr3_pb2.Instr
-Payload = payload_ddr3_pb2.Payload
+import payload_ddr4_pb2
+
+Timing = payload_ddr4_pb2.Timing
+Opcode = payload_ddr4_pb2.Opcode
+Instr = payload_ddr4_pb2.Instr
+Payload = payload_ddr4_pb2.Payload
 
 def VerifyInstr(ip: int, instr: Instr) -> bool:
   if instr.HasField('mem'):
@@ -15,6 +16,10 @@ def VerifyInstr(ip: int, instr: Instr) -> bool:
     if not (0 < mem.timeslice < (1 << Instr.MemInstr.Bits.TIMESLICE)):
       return False
     if mem.rank != 0:  # TODO: Add multi-rank support.
+      return False
+    if mem.stack != 0:  # TODO: Add die-stack support.
+      return False
+    if mem.bank_group >= 1 << Instr.MemInstr.Bits.BANK_GROUP:
       return False
     if mem.bank >= 1 << Instr.MemInstr.Bits.BANK:
       return False
@@ -47,11 +52,7 @@ def VerifyInstr(ip: int, instr: Instr) -> bool:
 class Rank:
   def __init__(self, timing: Timing):
     self.parameters = {
-        Opcode.RD: {
-            Opcode.RD: timing.ccd,
-        },
         Opcode.ACT: {
-            Opcode.ACT: timing.rrd,
             Opcode.REF: math.inf,
         },
         Opcode.PRE: {
@@ -69,9 +70,8 @@ class Rank:
     self.prev_acts = collections.deque(maxlen = 4)
     self.faw = timing.faw
 
-    # Banks.
-    self.banks = [
-        Bank(timing) for _ in range(1 << Instr.MemInstr.Bits.BANK)
+    self.bank_groups = [
+        BankGroup(timing) for _ in range(1 << Instr.MemInstr.Bits.BANK_GROUP)
     ]
 
   def Execute(self, tick: int, instr: Instr.MemInstr) -> bool:
@@ -90,8 +90,13 @@ class Rank:
           return False
       self.prev_acts.append(tick)
 
-    if not self.banks[instr.bank].Execute(tick, instr):
+    if not self.bank_groups[instr.bank_group].Execute(tick, instr):
       return False
+    for i, bank_group in enumerate(self.bank_groups):
+      if i == instr.bank_group:
+        continue
+      # Update all other bank groups.
+      bank_group.Update(tick, instr)
 
     for opcode, parameter in self.parameters.get(instr.opcode, {}).items():
       if self.next_tick[opcode] == math.inf:
@@ -100,6 +105,38 @@ class Rank:
         self.next_tick[opcode] = tick + parameter
     return True
 
+class BankGroup:
+  def __init__(self, timing: Timing):
+    self.parameters = {
+        Opcode.RD: {
+            Opcode.RD: [timing.ccd_l, timing.ccd_s]
+        },
+        Opcode.ACT: {
+            Opcode.ACT: [timing.rrd_l, timing.rrd_s]
+        },
+    }
+    self.next_tick = {Opcode.RD: 0, Opcode.ACT: 0}
+
+    self.banks = [
+        Bank(timing) for _ in range(1 << Instr.MemInstr.Bits.BANK)
+    ]
+
+  def Execute(self, tick: int, instr: Instr.MemInstr) -> bool:
+    if tick < self.next_tick.get(instr.opcode, 0):
+      print('Bank group timing violation for {}: {} < {}'.format(
+          Opcode.Name(instr.opcode), tick, self.next_tick[instr.opcode]))
+      return False
+
+    if not self.banks[instr.bank].Execute(tick, instr):
+      return False
+
+    for opcode, parameter in self.parameters.get(instr.opcode, {}).items():
+      self.next_tick[opcode] = tick + parameter[0]
+    return True
+
+  def Update(self, tick: int, instr: Instr.MemInstr):
+    for opcode, parameter in self.parameters.get(instr.opcode, {}).items():
+      self.next_tick[opcode] = tick + parameter[1]
 
 class Bank:
   def __init__(self, timing: Timing):
