+++ rowhammer_tester/payload/payloads/ddr3/activate_open_bank.textproto
