+++ rowhammer_tester/payload/payloads/ddr3/read_closed_bank.textproto
