+++ rowhammer_tester/payload/payloads/ddr3/refresh_open_bank.textproto
