+++ rowhammer_tester/payload/payloads/ddr4/activate_open_bank.textproto
