+++ rowhammer_tester/payload/payloads/ddr4/hammer.textproto
