+++ rowhammer_tester/payload/payloads/ddr4/read_closed_bank.textproto
