+++ rowhammer_tester/payload/payloads/ddr4/refresh.textproto
