+++ rowhammer_tester/payload/payloads/ddr4/refresh_open_bank.textproto
