--- rowhammer_tester/scripts/payloads/ddr3/hammer.textproto
@@ -1,30 +1,0 @@
-# payload_ddr3.proto.Payload
-timing: {
-  tick_ps: 1000
-  ras: 32
-  rp: 10
-  rfc: 103
-  rcd: 10
-  ccd: 4
-  rtp: 8
-  rrd: 5
-  faw: 26
-}
-instr: {
-  mem: {
-    opcode: ACT
-    timeslice: 32
-  }
-}
-instr: {
-  mem: {
-    opcode: PRE
-    timeslice: 9
-  }
-}
-instr: {
-  jmp: {
-    offset: 2
-    count: 99
-  }
-}
